pwcli: cmd_commit(): remove skip functionality entirely
Commit individually should not have the 'Skip' option, only Commit/Edit/aBort.
Skip option complicates the logic too much and it's possible to skip commits
with commit arguments already (like skip patch 8 with 'commit 5-7,9-10'). The
philosophy is that the patchset is handled as a whole and all patches get the
same result. Later, if really needed, we could reconsider adding option to skip
patches, for example if they fail to apply.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>