pwcli: add cache for patchwork patches
To reduce the need to query the patchwork server and make UI more responsive
add a cache for patches. The caches are per project and are not currently
updated ever, the user needs to restart pwcli to update the caches. Maybe a
refresh command would be good to have, or a timeout after which the caches are
automatically refreshed?

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>