pwcli: add cache for stg commits
For me 'stg show foo-patch-name' can take 0.6 seconds. If there are over 15
patches in pending branch it can take almost 10 seconds to update all of them.
To remove this unnecessary latency a cache for stg commit information. The
downside is that now it's not checked if the cache information is still valid,
that should be still added to avoid accidental use of old versions.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>