pwcli: create_patchlist_as_string(): sort patches before showing them
In commit 56d96a900492 ("pwcli: add cache for patchwork patches") we don't
respect the order in which the patchwork servers returns us the patches. For
some reason this was not noticed with cmdtests, only when I started to use it
with a real patchwork server.

Fix that by sorting patches based on date and subject field. Also try to get
the patch index and use that in sorting.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>