stubslib: convert GitRepository to use pickle
Now more ugly file parsing, instead have everything in GitRepository class
which is just stored to a file. This way it's a lot easier to extend it with
new features.

For this to work had to do quite a few changes to various test scripts. Tried
to keep the test output minimal, but in two cases I didn't see the point of
trying to keep the outputs same as before.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>