pwcli: list: make active states as default
Instead of making the review state as default if no state is defined, use
active active states as default. This makes it more intuitive to use the
results.

Also now pending branch patches are appended to the results, instead of
replacing them when requesting review state patches. This makes it easier to
extend the list command in followup patches.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>