pwcli: add stg list command
This is for printing all patches in the pending branch without any filters
which is useful when you want to see what already applied patches there are.

Other stg subcommands might be added later, like 'stg delete 2-5', 'stg rebase' and
so on to make it faster to manage the pending branch.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>