pwcli: support accessing using patchwork ids with style #12345
This way it's possible to access a patch directly using it's patchwork server
id. Multiple ids are possible to enter individually with format '#12345,#54321'.

Ranges (#100-#110) are not supported as patches from same series do not
necessarily have consecutive ids. This is visible especially with
patchwork.kernel.org which has lots of different projects.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>