pwcli: implement download command
A screenshot tells it all:

master@pwcli-stub-M_je2e > download 1-3
3 patches written to 1030.mbox
master@pwcli-stub-M_je2e >

Do note that I wrote this patch on the plane and I don't have access to
Internet because the antenna is broken, sigh. So I haven't had a chance to test
this with a real patchwork server yet, only with a stub.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>