cmdtests: update output based on changes in REST API
Now that both pwcli and Patchwork stub are converted to REST API the automated
tests can be updated. I tried to make the changes minimal but there are few
changes in the output listed below.

-User          : test (7477, 30052, 118371)
+User          : test

At the moment we don't use user ids for anything so I removed the ids altogether.

-http://localhost:8105/1001/
+http://localhost/fixme/12345

Now we get the URL from the server and the stub currently has it hardcoded.
It's important to the URL in the server as in future releases patchwork starts
using Message-Id in the link.

-CommitRef:
+CommitRef: None

Didn't bother to fix this.

-command failed: Patchwork id 777777 not found
+command failed: 500 Server Error: INTERNAL SERVER ERROR for url: http://localhost:8105/api/1.1/patches/777777/

Didn't bother to cleanup error handling related to this either, maybe in the
future.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>