pwcli: adapt list output based on the screen width
The name of patch is variable, in smaller screens we just don't use it that much.

I tried to minimise changes to cmdtest output by changing the dimensions
(dimensions=(24,109) for pexpect.spawn() but that again caused the old problem
prompt of failing prompt waits. But with 'git diff -w' the test output changes
don't look that bad.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>