pwcli: list: fix spacing for width
I had counted WIDTH_INDEX incorrectly, it's actually 6 characters and not 5.
This caused also the mysterious +1 in WIDTH_SEPARATORS :)

Noticed while trying to align cover letter names correctly with list command.
The test output when show_indexes is False changes one char to the right due to
the new spacing.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>