pwcli: list: show age instead of date
For a maintainer it's more important to know how old the patch, not the date.
So show the age (10h, 6d, 3m etc) instead of the date submitted. This is also
makes the output more compact. While at it also reserve space for future
statics (Acked-by, comments etc) in the list output.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>