pwcli: send_email: create Message-ID for every email
I got a report that the reporter was seeing mails from pwcli duplicated but
with different Message-ID, this is most likely due to problems with
smtp.codeaurora.org I'm using. Workaround this creating the Message-ID yourself
before contacting the SMTP server.

To not show random output in cmdtests, which would fail the tests, add
PWCLI_HARDCODE_MESSAGE_ID environment variable to consistent output for tests.

Signed-off-by: Kalle Valo <kvalo@codeaurora.org>