pwcli: git: replace slow filter-branch with commit --amend
I updated my git to 2.30.2 and suddenly commit command started to take ~10
seconds per patch. The reason was git filter-branch, in earlier version it was
fast but now super slow. The git documentation now recommends avoid using it.

There's a simple solution to this: use git commit --amend to modify the commit
log. It's fast and easy to use.

Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>