stubs: workaround for readline problem
In Debian 12 I noticed that run_stub was not working correctly, something was
eating most of the keyboard input and typing commands was very difficult, you
had to press the same character several times before it went through. I
couldn't find the root cause (this has been working fine earlier Debian
versions) but a workaround is to disable readline. So disable readline when
using run_stub or cmdtests.

Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>