stubs: cmdtestlib: set dimensions to (100, 121)
With Ubuntu 22.04 in github workflow the output from pexpect was different:

-Accepted/Under review/Changes requested/Rejected/New/Deferred/Superseded/aWaiting upstream/not aPplicable/rFc/aBort? a
+<rseded/aWaiting upstream/not aPplicable/rFc/aBort? a

I do not know why this is happening, I guess something has changed one of
Python libraries or something. I found a workaround to explicitly set columns
in dimensions 121 and then the output with both Debian 12 and Ubuntu 22.04
match. If columns are 120 they don't match anymore, I do not know why.

The only changes in test output are the longer lines, no functional changes.

Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>